Add load prop to Map
This enables the caller to have more control over how and when the mapkit resources are loaded. Instead of supporting more API sruface area in mapkit-react library I thought it made sense to just hand over control at the load point as a "advanced use" option. This could lead to foot-guns if the caller doesn't load the necessary resources and then tries to use a component from the lib that depends on the missing resource. In this case the error messages should be clear enough (mapkit says the resrouce needs to be loaded when using associated APIs). Needed this flexibility to support my own desires for optimizing map loading performance https://developer.apple.com/documentation/mapkitjs/loading_the_latest_version_of_mapkit_js#3331749.